+++ tests/pip_test_unified.html
@@ -1,0 +1,271 @@
+<!DOCTYPE html>
+<html lang="ja">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>PiP Test Page - macOS Spaces Compatible</title>
+    <style>
+      body {
+        font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
+          sans-serif;
+        margin: 0;
+        padding: 20px;
+        background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
+        color: white;
+        min-height: 100vh;
+      }
+
+      .container {
+        max-width: 1200px;
+        margin: 0 auto;
+        padding: 20px;
+      }
+
+      h1 {
+        text-align: center;
+        font-size: 2.5em;
+        margin-bottom: 10px;
+        text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
+      }
+
+      .subtitle {
+        text-align: center;
+        font-size: 1.2em;
+        margin-bottom: 40px;
+        opacity: 0.9;
+      }
+
+      .section {
+        background: rgba(255, 255, 255, 0.1);
+        border-radius: 15px;
+        padding: 30px;
+        margin-bottom: 30px;
+        backdrop-filter: blur(10px);
+        border: 1px solid rgba(255, 255, 255, 0.2);
+      }
+
+      .section h2 {
+        color: #fff;
+        margin-bottom: 20px;
+        font-size: 1.8em;
+      }
+
+      .image-grid {
+        display: grid;
+        grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+        gap: 20px;
+        margin-bottom: 20px;
+      }
+
+      .image-item {
+        position: relative;
+        border-radius: 10px;
+        overflow: hidden;
+        box-shadow: 0 8px 25px rgba(0, 0, 0, 0.3);
+        transition: transform 0.3s ease;
+      }
+
+      .image-item:hover {
+        transform: translateY(-5px);
+      }
+
+      .image-item img {
+        width: 100%;
+        height: 200px;
+        object-fit: cover;
+        display: block;
+      }
+
+      .image-caption {
+        position: absolute;
+        bottom: 0;
+        left: 0;
+        right: 0;
+        background: linear-gradient(transparent, rgba(0, 0, 0, 0.8));
+        color: white;
+        padding: 20px 15px 15px;
+        font-size: 14px;
+      }
+
+      .instructions {
+        background: rgba(255, 255, 255, 0.15);
+        border-radius: 10px;
+        padding: 20px;
+        margin-bottom: 20px;
+        border-left: 4px solid #007acc;
+      }
+
+      .shortcut {
+        font-family: "Monaco", monospace;
+        background: rgba(0, 0, 0, 0.3);
+        padding: 4px 8px;
+        border-radius: 4px;
+        font-weight: bold;
+        color: #ffd700;
+      }
+
+      .feature-list {
+        list-style: none;
+        padding: 0;
+      }
+
+      .feature-list li {
+        padding: 10px 0;
+        border-bottom: 1px solid rgba(255, 255, 255, 0.1);
+      }
+
+      .feature-list li:last-child {
+        border-bottom: none;
+      }
+
+      .feature-list li::before {
+        content: "✓";
+        color: #4caf50;
+        font-weight: bold;
+        margin-right: 10px;
+      }
+    </style>
+  </head>
+  <body>
+    <div class="container">
+      <h1>🖼️ PiP Test Page</h1>
+      <p class="subtitle">macOS Spaces Compatible Picture-in-Picture Testing</p>
+
+      <div class="section">
+        <h2>📖 How to Use</h2>
+        <div class="instructions">
+          <p><strong>キーボードショートカット:</strong></p>
+          <p>
+            画像をPicture-in-Pictureで表示:
+            <span class="shortcut">Ctrl+Alt+I</span>
+          </p>
+          <p>
+            すべてのPiPウィンドウを閉じる:
+            <span class="shortcut">Ctrl+Alt+X</span>
+          </p>
+        </div>
+
+        <h3>📋 Features</h3>
+        <ul class="feature-list">
+          <li>macOS Spacesを移動してもPiPウィンドウが表示され続ける</li>
+          <li>NSWindowのcollectionBehaviorを使用した独自PiP実装</li>
+          <li>画像のPicture-in-Picture表示</li>
+          <li>ドラッグ&ドロップでウィンドウ移動可能</li>
+          <li>ダブルクリックまたは×ボタンで閉じる</li>
+        </ul>
+      </div>
+
+      <div class="section">
+        <h2>🖼️ Test Images</h2>
+        <p>
+          以下の画像でPicture-in-Picture機能をテストしてください。<span
+            class="shortcut"
+            >Ctrl+Alt+I</span
+          >を押すと、適切な画像がPiPウィンドウで表示されます。
+        </p>
+
+        <div class="image-grid">
+          <div class="image-item">
+            <img
+              src="https://picsum.photos/400/300?random=1"
+              alt="Beautiful Landscape"
+            />
+            <div class="image-caption">
+              Beautiful Landscape - Perfect for PiP testing
+            </div>
+          </div>
+
+          <div class="image-item">
+            <img
+              src="https://picsum.photos/400/300?random=2"
+              alt="City Architecture"
+            />
+            <div class="image-caption">Modern Architecture - Test image #2</div>
+          </div>
+
+          <div class="image-item">
+            <img
+              src="https://picsum.photos/400/300?random=3"
+              alt="Nature Scene"
+            />
+            <div class="image-caption">
+              Nature Scene - High quality test image
+            </div>
+          </div>
+
+          <div class="image-item">
+            <img
+              src="https://picsum.photos/400/300?random=4"
+              alt="Abstract Art"
+            />
+            <div class="image-caption">
+              Abstract Art - Colorful test content
+            </div>
+          </div>
+        </div>
+      </div>
+
+      <div class="section">
+        <h2>🧪 Testing Instructions</h2>
+        <div class="instructions">
+          <h3>Basic Testing:</h3>
+          <ol style="line-height: 1.8">
+            <li>
+              ページを開いた状態で
+              <span class="shortcut">Ctrl+Alt+I</span> を押す
+            </li>
+            <li>PiPウィンドウが表示されることを確認</li>
+            <li>PiPウィンドウをドラッグして移動できることを確認</li>
+            <li>×ボタンまたはダブルクリックでウィンドウが閉じることを確認</li>
+          </ol>
+
+          <h3>macOS Spaces Testing:</h3>
+          <ol style="line-height: 1.8">
+            <li>
+              PiPウィンドウを表示した状態でSpacesを切り替え（Mission
+              Control使用）
+            </li>
+            <li>別のワークスペースでもPiPウィンドウが表示されることを確認</li>
+            <li>
+              複数のワークスペース間を移動してもPiPウィンドウが常に表示されることを確認
+            </li>
+          </ol>
+        </div>
+      </div>
+
+      <div class="section">
+        <h2>🔧 Technical Details</h2>
+        <p>この実装では以下の技術を使用しています：</p>
+        <ul class="feature-list">
+          <li>Qt Framework with macOS-specific Objective-C++ integration</li>
+          <li>NSWindow collectionBehavior for Spaces compatibility</li>
+          <li>Custom PiP window implementation (not HTML5 video PiP API)</li>
+          <li>JavaScript injection for image detection and extraction</li>
+          <li>QWebEngine integration for seamless browser experience</li>
+        </ul>
+      </div>
+    </div>
+
+    <script>
+      // PiP機能テスト用のJavaScript
+      console.log("PiP Test Page loaded");
+
+      // 画像クリック時のフィードバック
+      document.querySelectorAll(".image-item img").forEach((img) => {
+        img.addEventListener("click", function () {
+          console.log("Image clicked:", this.alt);
+          // 簡単な視覚的フィードバック
+          this.style.transform = "scale(0.95)";
+          setTimeout(() => {
+            this.style.transform = "scale(1)";
+          }, 150);
+        });
+      });
+
+      // PiP関連のイベントリスナー
+      document.addEventListener("pipImageSelected", function (event) {
+        console.log("PiP Image Selected:", event.detail);
+      });
+    </script>
+  </body>
+</html>
